Refactoring and optimization
