+++ flow-server/src/test/java/com/vaadin/tests/server/component/FlowClassesSerializableTest.java
@@ -1,0 +1,7 @@
+package com.vaadin.tests.server.component;
+
+import com.vaadin.flow.testutil.ClassesSerializableTest;
+
+public class FlowClassesSerializableTest extends ClassesSerializableTest
+{
+}
