+++ pom.xml
@@ -23,6 +23,7 @@
         <module>flow-server-production-mode</module>
         <module>flow-components-parent</module>
         <module>flow-maven-plugin</module>
+        <module>flow-test-generic</module>
         <module>build-tools</module>
     </modules>
